Fix problem with braces messing up syntax highlighting
As reported in #10, unclosed braces in templates can mess up the
syntax highlighting of the following code. It seems the problem is with
the `perlBraces` highlight region in `syntax/perl.vim` which uses the
`extend` argument. This means Perl highlight region extends beyond the
end of the template tag, messing the up the highlighting.

Excluding the `perlBraces` seems to fix this problem. I'm not sure if
there still are some cases where Perl code can cause the highlight
region to be extended.